Kod/pyproject.toml
[tool.poetry]
name = "kod"
version = "0.1.0"
description = ""
authors = ["philis-johan <philis@student.chalmers.se>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
jsonschema = "^4.17.3"
pandas = "^1.5.3"
matplotlib = "^3.7.1"
pygame = "^2.2.0"
seaborn = "^0.12.2"
pysrim = {git = "https://github.com/philis-johan/pysrim"}
pexpect = "^4.8.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.21.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
